Removed old uneccessary PL CUI. Cleaned up client code.
