+++ src/client/java/com/forgeessentials/client/core/ClientProxy.java
@@ -1,31 +1,34 @@
 package com.forgeessentials.client.core;
 
+import static com.forgeessentials.client.ForgeEssentialsClient.feclientlog;
+import static com.forgeessentials.client.ForgeEssentialsClient.netHandler;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.common.config.Configuration;
+
 import com.forgeessentials.client.ForgeEssentialsClient;
-import com.forgeessentials.client.cui.CUIPlayerLogger;
 import com.forgeessentials.client.cui.CUIRenderrer;
-import com.forgeessentials.client.cui.CUIRollback;
 import com.forgeessentials.client.network.C0PacketHandshake;
 import com.forgeessentials.client.network.C1PacketSelectionUpdate;
-import com.forgeessentials.client.network.C2PacketPlayerLogger;
-import com.forgeessentials.client.network.C3PacketRollback;
 import com.forgeessentials.client.network.C4PacketEconomy;
 import com.forgeessentials.client.network.C5PacketNoclip;
 import com.forgeessentials.client.util.DummyProxy;
 import com.forgeessentials.commons.VersionUtils;
+import com.forgeessentials.commons.selections.Selection;
+
 import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import cpw.mods.fml.common.eventhandler.SubscribeEvent;
+import cpw.mods.fml.common.gameevent.PlayerEvent.PlayerLoggedInEvent;
+import cpw.mods.fml.common.network.FMLNetworkEvent;
 import cpw.mods.fml.common.network.NetworkRegistry;
 import cpw.mods.fml.relauncher.Side;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.common.config.Configuration;
-
-import static com.forgeessentials.client.ForgeEssentialsClient.feclientlog;
-import static com.forgeessentials.client.ForgeEssentialsClient.netHandler;
 
 public class ClientProxy extends DummyProxy
 {
     private ClientConfig config;
+
+    private static Selection selection;
 
     @Override
     public void doPreInit(FMLPreInitializationEvent e)
@@ -39,8 +42,6 @@ public class ClientProxy extends DummyProxy
         netHandler = NetworkRegistry.INSTANCE.newSimpleChannel("forgeessentials");
         netHandler.registerMessage(C0PacketHandshake.class, C0PacketHandshake.class, 0, Side.SERVER);
         netHandler.registerMessage(C1PacketSelectionUpdate.class, C1PacketSelectionUpdate.class, 1, Side.CLIENT);
-        netHandler.registerMessage(C2PacketPlayerLogger.class, C2PacketPlayerLogger.class, 2, Side.CLIENT);
-        netHandler.registerMessage(C3PacketRollback.class, C3PacketRollback.class, 3, Side.CLIENT);
         netHandler.registerMessage(C4PacketEconomy.class, C4PacketEconomy.class, 4, Side.CLIENT);
         netHandler.registerMessage(C5PacketNoclip.class, C5PacketNoclip.class, 5, Side.CLIENT);
     }
@@ -49,13 +50,46 @@ public class ClientProxy extends DummyProxy
     public void load(FMLInitializationEvent e)
     {
         super.load(e);
-
-        FMLCommonHandler.instance().bus().register(new ClientEventHandler());
+        FMLCommonHandler.instance().bus().register(this);
         if (ForgeEssentialsClient.allowCUI)
         {
             MinecraftForge.EVENT_BUS.register(new CUIRenderrer());
-            MinecraftForge.EVENT_BUS.register(new CUIPlayerLogger());
-            MinecraftForge.EVENT_BUS.register(new CUIRollback());
         }
     }
+
+    @SubscribeEvent
+    public void connectionOpened(FMLNetworkEvent.ClientConnectedToServerEvent e)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isClient())
+            selection = null;
+
+    }
+
+    @SubscribeEvent
+    public void connectionClosed(FMLNetworkEvent.ClientDisconnectionFromServerEvent e)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isClient())
+            selection = null;
+    }
+
+    @SubscribeEvent
+    public void onPlayerLogin(PlayerLoggedInEvent e)
+    {
+        if (ForgeEssentialsClient.instance.serverHasFE)
+        {
+            System.out.println("Dispatching FE handshake packet");
+            ForgeEssentialsClient.netHandler.sendToServer(new C0PacketHandshake());
+        }
+    }
+
+    public static Selection getSelection()
+    {
+        return selection;
+    }
+    
+    public static void setSelection(Selection sel)
+    {
+        selection = sel;
+    }
+    
 }
