--- src/client/java/com/forgeessentials/client/cui/CUIPlayerLogger.java
@@ -1,161 +1,0 @@
-package com.forgeessentials.client.cui;
-
-import com.forgeessentials.client.ForgeEssentialsClient;
-import com.forgeessentials.client.core.PlayerInfoClient;
-import com.forgeessentials.commons.selections.Point;
-import cpw.mods.fml.client.FMLClientHandler;
-import cpw.mods.fml.common.eventhandler.SubscribeEvent;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
-import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.client.renderer.entity.RenderManager;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.util.MathHelper;
-import net.minecraft.util.MovingObjectPosition;
-import net.minecraft.util.Vec3;
-import net.minecraftforge.client.event.RenderWorldLastEvent;
-import org.lwjgl.opengl.GL11;
-import org.lwjgl.opengl.GL12;
-
-@SideOnly(value = Side.CLIENT)
-public class CUIPlayerLogger {
-    @SubscribeEvent
-    public void render(RenderWorldLastEvent event)
-    {
-        EntityPlayer player = FMLClientHandler.instance().getClient().thePlayer;
-        PlayerInfoClient info = ForgeEssentialsClient.info;
-
-        if (player == null || info == null || !info.playerLogger)
-        {
-            return;
-        }
-
-        Point p = getPoint(player);
-
-        if (p == null)
-        {
-            return;
-        }
-
-        GL11.glPushMatrix();
-        GL11.glEnable(GL12.GL_RESCALE_NORMAL);
-        GL11.glDisable(GL11.GL_DEPTH_TEST);
-        GL11.glDisable(GL11.GL_TEXTURE_2D);
-        Tessellator tess = Tessellator.instance;
-        Tessellator.renderingWorldRenderer = false;
-
-        GL11.glTranslated(p.x - RenderManager.renderPosX, p.y + 1 - RenderManager.renderPosY, p.z - RenderManager.renderPosZ);
-        GL11.glScalef(1.0F, -1.0F, -1.0F);
-        GL11.glColor3f(255, 145, 0);
-        renderBlockBox(tess);
-
-        GL11.glDisable(GL12.GL_RESCALE_NORMAL);
-        GL11.glEnable(GL11.GL_DEPTH_TEST);
-        GL11.glEnable(GL11.GL_TEXTURE_2D);
-        GL11.glPopMatrix();
-    }
-
-    private Point getPoint(EntityPlayer player)
-    {
-        float var4 = 1.0F;
-        float var5 = player.prevRotationPitch + (player.rotationPitch - player.prevRotationPitch) * var4;
-        float var6 = player.prevRotationYaw + (player.rotationYaw - player.prevRotationYaw) * var4;
-        double var7 = player.prevPosX + (player.posX - player.prevPosX) * var4;
-        double var9 = player.prevPosY + (player.posY - player.prevPosY) * var4 + 1.62D - player.yOffset;
-        double var11 = player.prevPosZ + (player.posZ - player.prevPosZ) * var4;
-        Vec3 var13 = Vec3.createVectorHelper(var7, var9, var11);
-        float var14 = MathHelper.cos(-var6 * 0.017453292F - (float) Math.PI);
-        float var15 = MathHelper.sin(-var6 * 0.017453292F - (float) Math.PI);
-        float var16 = -MathHelper.cos(-var5 * 0.017453292F);
-        float var17 = MathHelper.sin(-var5 * 0.017453292F);
-        float var18 = var15 * var16;
-        float var20 = var14 * var16;
-        double var21 = 500D;
-        Vec3 var23 = var13.addVector(var18 * var21, var17 * var21, var20 * var21);
-
-        MovingObjectPosition mo = player.worldObj.rayTraceBlocks(var13, var23);
-
-        if (mo == null)
-        {
-            return null;
-        }
-
-        Point p = new Point(mo.blockX, mo.blockY, mo.blockZ);
-
-        if (!player.isSneaking())
-        {
-            return p;
-        }
-
-        switch (mo.sideHit)
-        {
-        case 0:
-            p.y--;
-            break;
-        case 1:
-            p.y++;
-            break;
-        case 2:
-            p.z--;
-            break;
-        case 3:
-            p.z++;
-            break;
-        case 4:
-            p.x--;
-            break;
-        case 5:
-            p.x++;
-            break;
-        }
-
-        return p;
-    }
-
-    /**
-     * must be translated to proper point before calling
-     */
-    private void renderBlockBox(Tessellator tess)
-    {
-        tess.startDrawing(GL11.GL_LINES);
-
-        // FRONT
-        tess.addVertex(0, 0, 0);
-        tess.addVertex(0, 1, 0);
-
-        tess.addVertex(0, 1, 0);
-        tess.addVertex(1, 1, 0);
-
-        tess.addVertex(1, 1, 0);
-        tess.addVertex(1, 0, 0);
-
-        tess.addVertex(1, 0, 0);
-        tess.addVertex(0, 0, 0);
-
-        // BACK
-        tess.addVertex(0, 0, -1);
-        tess.addVertex(0, 1, -1);
-        tess.addVertex(0, 0, -1);
-        tess.addVertex(1, 0, -1);
-        tess.addVertex(1, 0, -1);
-        tess.addVertex(1, 1, -1);
-        tess.addVertex(0, 1, -1);
-        tess.addVertex(1, 1, -1);
-
-        // betweens.
-        tess.addVertex(0, 0, 0);
-        tess.addVertex(0, 0, -1);
-
-        tess.addVertex(0, 1, 0);
-        tess.addVertex(0, 1, -1);
-
-        tess.addVertex(1, 0, 0);
-        tess.addVertex(1, 0, -1);
-
-        tess.addVertex(1, 1, 0);
-        tess.addVertex(1, 1, -1);
-
-        tess.draw();
-    }
-
-}
