+++ src/client/java/com/forgeessentials/client/gui/forge/FEConfigGUI.java
@@ -1,10 +1,12 @@
 package com.forgeessentials.client.gui.forge;
 
-import com.forgeessentials.client.core.ClientConfig;
-import cpw.mods.fml.client.config.GuiConfig;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraftforge.common.config.ConfigElement;
 import net.minecraftforge.common.config.Configuration;
+
+import com.forgeessentials.client.core.ClientConfig;
+
+import cpw.mods.fml.client.config.GuiConfig;
 
 public class FEConfigGUI extends GuiConfig {
 
@@ -14,4 +16,5 @@ public class FEConfigGUI extends GuiConfig {
                 new ConfigElement(ClientConfig.config.getCategory(Configuration.CATEGORY_GENERAL)).getChildElements(),
                 "TestMod", false, false, "FE Client Addon Config");
     }
+    
 }
