+++ src/main/java/com/forgeessentials/playerlogger/command/RollbackInfo.java
@@ -165,6 +165,7 @@ public class RollbackInfo {
         }
 
         @Override
+        @SuppressWarnings("deprecation")
         public void run()
         {
             rb.getTime().setSeconds(rb.getTime().getSeconds() - speed);
